backend/requirements.txt
# FastAPI + Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2,<5.0.0

# Pydantic
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Google APIs (for dual-write + legacy code)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
gspread>=5.12.0

# Clover / Shopify API
requests>=2.31.0
python-dateutil>=2.8.2

# Data processing
pytz>=2023.3

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
